PR Feedback: enhanced test + comment on PARENTS_AND_SPECS
