+++ lucille-core/src/test/resources/RunnerTest/allBadOtherParents.conf
@@ -29,5 +29,20 @@ runner {
   metricsLoggingLevel: "INFO"
 
   something: "abcdef"
-  somethingElse: "123456"
+}
+
+log {
+  something: true
+}
+
+publisher {
+  something: "10s"
+}
+
+kafka {
+  something: "abc"
+}
+
+worker {
+  something: "123"
 }
